requirements.txt
fastapi[all]==0.115.0
slowapi==0.1.9
pytest==8.3.2
httpx==0.27.0
uvicorn==0.32.0
sqlalchemy==2.0.36
psycopg[binary]
pydantic==2.9.2
alembic
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest
pydantic-settings==2.5.2
psycopg2-binary
